requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0